+++ packages/decap-cms-widget-object/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.1](https://github.com/decaporg/decap-cms/compare/decap-cms-widget-object@3.3.0...decap-cms-widget-object@3.3.1) (2025-01-30)
+
+### Bug Fixes
+
+- **ObjectControl:** hotfix nested object validation ([#7385](https://github.com/decaporg/decap-cms/issues/7385)) ([3f5461c](https://github.com/decaporg/decap-cms/commit/3f5461c4e854ce3b9a7fb08974e270d54f5fccff))
+
+# [3.3.0](https://github.com/decaporg/decap-cms/compare/decap-cms-widget-object@3.2.0...decap-cms-widget-object@3.3.0) (2025-01-29)
+
+### Features
+
+- visual editing (click-to-edit) ([#7374](https://github.com/decaporg/decap-cms/issues/7374)) ([989c2dd](https://github.com/decaporg/decap-cms/commit/989c2dd6ed80f69b572b8b73c4e37b5106ae04fb))
+
+# [3.2.0](https://github.com/decaporg/decap-cms/compare/decap-cms-widget-object@3.1.4...decap-cms-widget-object@3.2.0) (2025-01-15)
+
+### Bug Fixes
+
+- **object-widget:** highlight nested validation errors ([#7330](https://github.com/decaporg/decap-cms/issues/7330)) ([0bb2904](https://github.com/decaporg/decap-cms/commit/0bb290492218e710e113cadccc49cdc475328db4))
+
 ## [3.1.4](https://github.com/decaporg/decap-cms/compare/decap-cms-widget-object@3.1.3...decap-cms-widget-object@3.1.4) (2024-08-13)
 
 ### Reverts